Done edit user

+++ app/Helpers/Resource.php
@@ -7,8 +7,13 @@ use Intervention\Image\Laravel\Facades\Image;
 
 class Resource {
     public static function delete($dirName, $file){
-        $rs = Storage::delete($dirName.'/'.$file);
-        return ($rs) ? $rs : false;
+        $path = $dirName.'/'.$file;
+        if(Storage::exists($path)){
+            $rs = Storage::delete($dirName.'/'.$file);
+            return ($rs) ? $rs : false;
+        }else{
+            return true;
+        }
     }
 
     public static function uploadImage($dirName, $image, $type = 'default')
